requirements.txt
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
openai>=1.52.0
pdfplumber==0.11.0
pdf2image==1.17.0
python-docx==1.1.2
python-dotenv==1.0.1
pytest==8.2.0
pytest-mock==3.14.0
